--- Core/InOut/AllFood.txt
@@ -1,32 +1,0 @@
-    -Dinh dang-
-        ID
-        Name
-        Thoi diem an trong ngay (1\2\3\12\13\23\123)
-            Buoi 1(Buoi dau trong ngay) / Buoi 2(Buoi an thu 2 trong ngay) / Buoi 3
-        Ke`m Com(0\1)
-            Co phai la mon an kem voi com hay khong
-        Ngay chon cuoi
-        Lan chon
-        Calories
-
-1000001
-Com Suon
-123
-0
-0
-0
-565
-1000002
-Suon xao chua ngot
-23
-1
-0
-0
-370
-1000003
-Hu tieu
-13
-0
-0
-0
-650
