--- Core/InOut/DailyFood.txt
@@ -1,21 +1,0 @@
-Banh_Xeo
-Hu_Tieu_Kho^
-Bot_Chien
-Sushi
-Suon_xao_chua_ngot
-Bo_luc_lac
-Banh_mi
-Com_suon
-Banh_Uot
-Banh_Cuon
-Bun_thit_nuong
-Hu_tieu_xao
-Banh_mi_nuong_bo
-Mi_cay_kho
-Xoi
-Com_ga_xoi_mo
-Mi_tuong_den
-Banh_mi_nuong_muoi_ot
-Banh_mi_cham_sua
-Banh_mi_hap_thit_bam
-Mi_y_sot_kem_bo_bam
