--- Core/InOut/FoodAfterFilter.txt
@@ -1,21 +1,0 @@
-1000001
-Com Suon
-123
-0
-0
-0
-565
-1000002
-Suon xao chua ngot
-23
-1
-0
-0
-370
-1000003
-Hu tieu
-13
-0
-0
-0
-650
