--- Core/InOut/WeeklyLog.txt
@@ -1,106 +1,0 @@
-
-100006
-Banh_Xeo
-1
-03/07/2023
-315
-100007
-Hu_Tieu_Kho^
-1
-03/07/2023
-379
-100008
-Bot_Chien
-1
-03/07/2023
-340
-100012
-Sushi
-1
-03/07/2023
-300
-100017
-Suon_xao_chua_ngot
-1
-03/07/2023
-360
-100019
-Bo_luc_lac
-1
-03/07/2023
-350
-100020
-Banh_mi
-1
-03/07/2023
-264
-100001
-Com_suon
-1
-03/07/2023
-565
-100002
-Banh_Uot
-1
-03/07/2023
-590
-100003
-Banh_Cuon
-1
-03/07/2023
-590
-100014
-Bun_thit_nuong
-1
-03/07/2023
-556
-100015
-Hu_tieu_xao
-1
-03/07/2023
-646
-100027
-Banh_mi_nuong_bo
-1
-03/07/2023
-561
-100030
-Mi_cay_kho
-1
-03/07/2023
-545
-100005
-Xoi
-1
-03/07/2023
-500
-100011
-Com_ga_xoi_mo
-1
-03/07/2023
-500
-100018
-Mi_tuong_den
-1
-03/07/2023
-420
-100025
-Banh_mi_nuong_muoi_ot
-1
-03/07/2023
-415
-100026
-Banh_mi_cham_sua
-1
-03/07/2023
-415
-100028
-Banh_mi_hap_thit_bam
-1
-03/07/2023
-500
-100037
-Mi_y_sot_kem_bo_bam
-1
-03/07/2023
-450
