--- Core/InOut/conditionActivate.txt
@@ -1,2 +1,0 @@
-07/03/23
-0
