--- Core/InOut/heightWeight.txt
@@ -1,5 +1,0 @@
-0
-171
-71
-17
-1
