+++ Core/clearFoodLog.cpp
@@ -1,6 +1,5 @@
 #include <stdio.h>
- 
-char WeeklyLog[100] = "InOut\\WeeklyLog.txt";
+#define WeeklyLog "InOut\\WeeklyLog.txt"
 
 int main(){
     FILE *f = fopen(WeeklyLog,"w");
