+++ Core/clearTotalCalories.cpp
@@ -1,5 +1,5 @@
 #include <stdio.h>
-char TotalCalories[100] = "InOut\\TotalCalories.txt";
+#define TotalCalories "InOut\\TotalCalories.txt"
 
 int main(){
     FILE *f = fopen(TotalCalories,"w");
