+++ TheMenu.py
@@ -15,6 +15,8 @@ from kivy.uix.image import Image
 from kivy.uix.floatlayout import FloatLayout
 from kivy.core.window import Window
 import os
+
+
 path_to = os.getcwd()
 os.chdir(path_to)
